# docs/swagger.yaml
basePath: /
definitions:
  model.Order:
    properties:
      created_at:
        type: string
      delivered_at:
        type: string
      id:
        type: integer
      products:
        items:
          $ref: '#/definitions/model.Product'
        type: array
      user_id:
        type: integer
    type: object
  model.Product:
    properties:
      amount:
        type: integer
      category:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      image_url:
        type: string
      imageID:
        type: string
      order_id:
        type: integer
      price:
        type: number
      purchased_amount:
        type: integer
      related_products:
        items:
          $ref: '#/definitions/model.Product'
        type: array
      reviews:
        items:
          $ref: '#/definitions/model.Review'
        type: array
      tag:
        type: string
      title:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
      views:
        type: integer
    required:
    - amount
    - category
    - price
    - title
    type: object
  model.Review:
    properties:
      category:
        type: string
      created_at:
        type: string
      id:
        type: integer
      product_id:
        type: integer
      text:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
      username:
        type: string
    required:
    - category
    - text
    type: object
...
  v1.errorResponse:
    properties:
      message:
        type: string
    type: object
  v1.getOrdersResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/model.Order'
        type: array
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Simple market API
  title: Market API
  version: "1.0"
paths:
...
  /api/v1/order:
    post:
      operationId: create-order
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.getOrdersResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/v1.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create order
      tags:
      - order
  /api/v1/order/{orderId}:
    get:
      operationId: get-order
      parameters:
      - description: ID of order to get
        in: path
        name: orderId
        required: true
        type: integer
      - description: sort by
        enum:
        - views
        - price
        - created_at
        in: query
        name: sort_by
        type: string
      - description: sort order
        enum:
        - asc
        - desc
        in: query
        name: sort_order
        type: string
      - description: limit
        enum:
        - 10
        - 25
        - 50
        in: query
        name: limit
        type: integer
      - description: page
        in: query
        name: page
        type: integer
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Order'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/v1.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get order
      tags:
      - order
  /api/v1/orders:
    get:
      operationId: get-orders
      parameters:
      - description: sort by
        enum:
        - created_at
        in: query
        name: sort_by
        type: string
      - description: sort order
        enum:
        - asc
        - desc
        in: query
        name: sort_order
        type: string
      - description: limit
        enum:
        - 10
        - 25
        - 50
        in: query
        name: limit
        type: integer
      - description: page
        in: query
        name: page
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.getOrdersResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/v1.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get orders
      tags:
      - order
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"